5258: balance view page
