+++ src/App.js
@@ -31,7 +31,7 @@ import FeedbackModal from './components/FeedbackModal'
 import GDPRPage from './pages/GDPRPage/GDPRPage'
 import AssetsPage from './pages/assets/AssetsPage'
 import SignalFormPage from './ducks/Signals/SignalFormPage'
-import HistoricalBalancePage from './ducks/HistoricalBalance/HistoricalBalancePage'
+import HistoricalBalancePage from './ducks/HistoricalBalance/page/HistoricalBalancePage'
 import WordCloudPage from './components/WordCloud/WordCloudPage'
 import { getConsentUrl } from './utils/utils'
 import NewsBanner from './components/Banner/NewsBanner'
