+++ src/ducks/HistoricalBalance/page/HistoricalBalancePage.module.scss
@@ -1,18 +1,27 @@
+@import "~@santiment-network/ui/mixins.scss";
 @import 'variables.scss';
-@import './../../mixins.scss';
+
+.header {
+  display: flex;
+  width: 100%;
+}
 
 .title {
   margin-bottom: 1rem;
 
-  h1 {
-    width: 100%;
-    font-weight: 600;
-    font-style: normal;
-    line-height: 28px !important;
-    font-size: 24px !important;
-    color: var(--black);
-    margin-left: 0;
-  }
+  @include text('h3');
+
+  color: var(--mirage);
+}
+
+.newTrigger {
+  margin-left: auto;
+}
+
+.questionIcon {
+  fill: var(--casper);
+  cursor: pointer;
+  margin-left: 15px;
 }
 
 /* stylelint-disable */
