+++ src/ducks/Signals/common/queries.js
